+++ streamlit/component-template/template-vite/vite_vanilla_component/__init__.py
@@ -1,106 +1,63 @@
 import os
 import streamlit.components.v1 as components
 
-# Create a _RELEASE constant. We'll set this to False while we're developing
-# the component, and True when we're ready to package and distribute it.
-# (This is, of course, optional - there are innumerable ways to manage your
-# release process.)
 _RELEASE = False
 
-# Declare a Streamlit component. `declare_component` returns a function
-# that is used to create instances of the component. We're naming this
-# function "_component_func", with an underscore prefix, because we don't want
-# to expose it directly to users. Instead, we will create a custom wrapper
-# function, below, that will serve as our component's public API.
-
-# It's worth noting that this call to `declare_component` is the
-# *only thing* you need to do to create the binding between Streamlit and
-# your component frontend. Everything else we do in this file is simply a
-# best practice.
-
 if not _RELEASE:
-    _component_func = components.declare_component(
-        # We give the component a simple, descriptive name ("my_component"
-        # does not fit this bill, so please choose something better for your
-        # own component :)
-        "vite_vanilla_component",
-        # Pass `url` here to tell Streamlit that the component will be served
-        # by the local dev server that you run via `npm run start`.
-        # (This is useful while your component is in development.)
-        url="http://localhost:3001",
-    )
+  _component_func = components.declare_component(
+    "vite_vanilla_component",
+    url="http://localhost:3000", # vite dev server port
+  )
 else:
-    # When we're distributing a production version of the component, we'll
-    # replace the `url` param with `path`, and point it to to the component's
-    # build directory:
-    parent_dir = os.path.dirname(os.path.abspath(__file__))
-    build_dir = os.path.join(parent_dir, "frontend/dist")
-    _component_func = components.declare_component("vite_vanilla_component", path=build_dir)
+  parent_dir = os.path.dirname(os.path.abspath(__file__))
+  build_dir = os.path.join(parent_dir, "frontend/dist")
+  _component_func = components.declare_component("vite_vanilla_component", path=build_dir)
 
+def my_component(name, config, nodes, edges, key=None):
+  component_value = _component_func(name=name, config=config, nodes=nodes, edges=edges, key=key, default=0)
+  return component_value
 
-# Create a wrapper function for the component. This is an optional
-# best practice - we could simply expose the component function returned by
-# `declare_component` and call it done. The wrapper allows us to customize
-# our component's API: we can pre-process its input args, post-process its
-# output value, and add a docstring for users.
-def my_component(name, key=None):
-    """Create a new instance of "vite_vanilla_component".
-
-    Parameters
-    ----------
-    name: str
-        The name of the thing we're saying hello to. The component will display
-        the text "Hello, {name}!"
-    key: str or None
-        An optional key that uniquely identifies this component. If this is
-        None, and the component's arguments are changed, the component will
-        be re-mounted in the Streamlit frontend and lose its current state.
-
-    Returns
-    -------
-    int
-        The number of times the component's "Click Me" button has been clicked.
-        (This is the value passed to `Streamlit.setComponentValue` on the
-        frontend.)
-
-    """
-    # Call through to our private component function. Arguments we pass here
-    # will be sent to the frontend, where they'll be available in an "args"
-    # dictionary.
-    #
-    # "default" is a special argument that specifies the initial return
-    # value of the component before the user has interacted with it.
-    component_value = _component_func(name=name, key=key, default=0)
-
-    # We could modify the value returned from the component if we wanted.
-    # There's no need to do this in our simple example - but it's an option.
-    return component_value
-
-
-# Add some test code to play with the component while it's in development.
-# During development, we can run this just as we would any other Streamlit
-# app: `$ streamlit run my_component/__init__.py`
 if not _RELEASE:
-    import streamlit as st
+  import streamlit as st
 
-    st.subheader("Component with constant args")
+  config = {
+    "container": "container",
+    "fitCenter": True,
+    "linkCenter": True,
+    "defaultNode": {
+      "type": "circle",
+      "size": [40],
+      "color": "#5B8FF9",
+      "style": { "fill": "#9EC9FF", "lineWidth": 3 },
+      "labelCfg": { "style": { "fill": "#000", "fontSize": 14 } }
+    },
+    "defaultEdge": {
+      "type": "quadratic",
+      "labelCfg": { "autoRotate": True, },
+    },
+    "modes": {
+      "default": ["drag-canvas", "drag-node"],
+    },
+    "nodeStateStyles": {
+      "hover": { "fillOpacity": 0.8 },
+      "selected": { "lineWidth": 5 }
+    }
+  }
 
-    # Create an instance of our component with a constant `name` arg, and
-    # print its output value.
-    num_clicks = my_component("World")
-    st.markdown("You've clicked %s times!" % int(num_clicks))
+  nodes = [
+    { "id": "node1", "x": 50, "y": 350, "label": "A", },
+    { "id": "node2", "x": 250, "y": 150, "label": "B", },
+    { "id": "node3", "x": 450, "y": 350, "label": "C", },
+  ]
+  edges = []
 
-    st.markdown("---")
-    st.subheader("Component with variable args")
+  for x in range(8):
+    edges.append({ "source": "node1", "target": "node2", "label": f'{x}th edge of A-B', })
+  for x in range(5):
+    edges.append({ "source": "node2", "target": "node3", "label": f'{x}th edge of B-C', })
 
-    # Create a second instance of our component whose `name` arg will vary
-    # based on a text_input widget.
-    #
-    # We use the special "key" argument to assign a fixed identity to this
-    # component instance. By default, when a component's arguments change,
-    # it is considered a new instance and will be re-mounted on the frontend
-    # and lose its current state. In this case, we want to vary the component's
-    # "name" argument without having it get recreated.
-    name_input = st.text_input("Enter a name", value="Streamlit")
-    num_clicks = my_component(name_input, key="foo")
-    st.markdown("You've clicked %s times!" % int(num_clicks))
+  st.subheader("Component Test")
+  rv = my_component(name="NameViteVanilla", config=config, nodes=nodes, edges=edges)
+  st.write(rv)
+  # st.markdown("You've clicked %s times!" % int(rv['numClicks']))
+  # st.markdown(f'Selected Node Is: {rv['selectedNode']}')