+++ streamlit/component-template/template-vite/vite_vanilla_component/frontend/main.js
@@ -1,75 +1,84 @@
 import { Streamlit } from "streamlit-component-lib"
+import G6 from '@antv/g6'
 
-// Add text and a button to the DOM. (You could also add these directly
-// to index.html.)
+let selectedNode = null
+let graph = null
+
 const span = document.body.appendChild(document.createElement("span"))
 const textNode = span.appendChild(document.createTextNode(""))
 const button = span.appendChild(document.createElement("button"))
 button.textContent = "Click Me!"
 
-// Add a click handler to our button. It will send data back to Streamlit.
 let numClicks = 0
 let isFocused = false
+button.onfocus = function() { isFocused = true }
+button.onblur = function() { isFocused = false }
 button.onclick = function() {
-  // Increment numClicks, and pass the new value back to
-  // Streamlit via `Streamlit.setComponentValue`.
   numClicks += 1
-  Streamlit.setComponentValue(numClicks)
+  Streamlit.setComponentValue({ numClicks, selectedNode })
 }
 
-button.onfocus = function() {
-  isFocused = true
-}
-
-button.onblur = function() {
-  isFocused = false
-}
-
-/**
- * The component's render function. This will be called immediately after
- * the component is initially loaded, and then again every time the
- * component gets new data from Python.
- */
 function onRender(event) {
-  // Get the RenderData from the event
+  console.log('RENDER EVENT', Date.now()) //  current status... of numClicks and selectedNode...
   const data = event.detail
-
-  // Maintain compatibility with older versions of Streamlit that don't send
-  // a theme object.
   if (data.theme) {
-    // Use CSS vars to style our button border. Alternatively, the theme style
-    // is defined in the data.theme object.
     const borderStyling = `1px solid var(${
       isFocused ? "--primary-color" : "gray"
     })`
     button.style.border = borderStyling
     button.style.outline = borderStyling
   }
-
-  // Disable our button if necessary.
   button.disabled = data.disabled
-
-  // RenderData.args is the JSON dictionary of arguments sent from the
-  // Python script.
   let name = data.args["name"]
-
-  // Show "Hello, name!" with a non-breaking space afterwards.
   textNode.textContent = `Hello, ${name}! ` + String.fromCharCode(160)
 
-  // We tell Streamlit to update our frameHeight after each render event, in
-  // case it has changed. (This isn't strictly necessary for the example
-  // because our height stays fixed, but this is a low-cost function, so
-  // there's no harm in doing it redundantly.)
+  let nodes = data.args["nodes"]
+  let edges = data.args["edges"]
+  const gdata = {
+    edges: [],
+    nodes: []
+  };
+  gdata.nodes = nodes
+  gdata.edges = edges  
+  G6.Util.processParallelEdges(gdata.edges);
+
+  let config = data.args["config"]
+  const container = document.body.appendChild(document.createElement("div"))
+  container.setAttribute('id', 'container');
+
+  if (!graph) {
+    config.width = document.getElementById('container').scrollWidth;
+    config.height = document.getElementById('container').scrollHeight || 500;
+  
+    graph = new G6.Graph(config)
+
+    // graph events handling
+    graph.on('edge:mouseenter', ({item}) => graph.setItemState(item, 'active', true))
+    graph.on('edge:mouseleave', ({item}) => graph.setItemState(item, 'active', false))
+    graph.on('edge:click', ({item}) => graph.setItemState(item, 'selected', true))
+    graph.on('node:click', (evt) => {
+      const { item } = evt
+      graph.getNodes().forEach((node) => graph.clearItemStates(node))
+      graph.setItemState(item, 'selected', true)
+      const { id, label } =  item._cfg.model
+      selectedNode = { id, label } // ${id} ${label}`
+    })
+    graph.on('canvas:click', (evt) => {
+      selectedNode = null
+      graph.getEdges().forEach((edge) => graph.clearItemStates(edge))
+      graph.getNodes().forEach((node) => graph.clearItemStates(node))
+    })
+  }
+  graph.data(gdata);
+  graph.render();
+  if (selectedNode) { // highlight node based on label
+    const foundNode = graph.getNodes().find((node) => node._cfg.model.label === selectedNode.label)
+    if (foundNode) graph.setItemState(foundNode, 'selected', true)
+  }
+
   Streamlit.setFrameHeight()
 }
 
-// Attach our `onRender` handler to Streamlit's render event.
 Streamlit.events.addEventListener(Streamlit.RENDER_EVENT, onRender)
-
-// Tell Streamlit we're ready to start receiving data. We won't get our
-// first RENDER_EVENT until we call this function.
 Streamlit.setComponentReady()
-
-// Finally, tell Streamlit to update our initial height. We omit the
-// `height` parameter here to have it default to our scrollHeight.
 Streamlit.setFrameHeight()